avm1: Moved start_drag from Avm1 root to activation module
